.gitignore: liberalize vim swapfiles selector
Sometimes, i.e., if you ignore a vim swapfile and start editing it, vim
appears to create a second swapfile which it names .<filename>.swo, in
addition to the already-existing .<filename>.swp file.  I'm guessing
that if you then kill vim and start editing a third time, it creates a
.<filename>.swn, and so on, backwards through the alphabet, so this
changes the .gitignore selector to ignore the last character.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>