LinkedList: tombstone feature for removal during iteration
When multiple items are removed during an ongoing node iteration, it
is possible that the node passed to nodeiter() has already been removed.
This becomes a problem if subsequent removals have occured because, since
the node is no longer in the list, it's .next may point somewhere that is
no longer the correct next node.

Therefore, mark each node as a "tombstone" as it is removed, and skip these
nodes when an iterator encounters them.

Also implement tests that detect this edge-case.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>