LinkedList: [ROBUST CLEAR 2/2]: Make :clear() robust
LinkedList:clear() has more-or-less exactly the same problem with
corruption of ongoing iterators.  To fix this, remember all the
ongoing unique iterations and, for each, the most recently iterated
node, if any.  When a clear() occurs.

Armed with this information, mark all subsequent nodes of all
ongoing iterators as tombstones whenever a clear() is issued.

Also implement corresponding test.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>